[outpolic-LSJ] 결제 화면 및 JS 수정 / 결제 controller 수정 detail : 필요한 데이터를 받기 위한 JS수정, input태그에 name과 id를 설정 / 상품 결제 controller 수정했씁니다

+++ outpolic/src/main/java/outpolic/user/payment/dto/SettlementDTO.java
@@ -15,7 +15,7 @@ public class SettlementDTO {
 	private int stlmCnt;
 	private String stlmPayType;
 	private BigDecimal stlmAmt;
-    private BigDecimal stlmUsedPoints;
+    private Integer stlmUsedPoints;
     private BigDecimal stlmFinalAmt;
     private String stlmPaymentInfo;
     private String stlmCardNm;
